Committing new files in /models adding two js files, also updated the seeds.sql, and moved the transporter variable into api-routes; deleted two topics and minimized check-box.
