+++ db/seeds.sql
@@ -1,19 +1,36 @@
 USE passport_demo;
 
-INSERT INTO topics (topic_name, t_condition)
-VALUES ("topic1", false),
-    ("topic2", false),
-    ("topic3", false),
-    ("topic4", false),
-    ("topic5", false),
-    ("topic6", false),
-    ("topic7", false);
+INSERT INTO html (html_name, html_condition)
+VALUES ("HTML Elements: headings, paragraphs and links", false),
+    ("HTML images, video and audio content", false),
+    ("Use HTML to solve problems", false),
+    ("Basic Construction of an HTML Page", false),
+    ("HTML Attributes Reference Guide", false);
+   
 ​​
-INSERT INTO activities (activity_name, a_condition)
-VALUES ("act1", false),
-    ("act2", false),
-    ("act3", false),
-    ("act4", false),
-    ("act5", false),
-    ("act6", false),
-    ("act7", false);
+INSERT INTO css (css_name, css_condition)
+VALUES ("CSS basic selectors and combinators", false),
+    ("CSS syntax and semantics", false),
+    ("Applying color to HTML using CSS", false),
+    ("Block formatting context concept", false),
+    ("Key concepts of Css: values, layout", false),
+   
+
+    INSERT INTO php (php_name, php_condition)
+VALUES ("Php basic syntax and langauge reference", false),
+    ("Php functions and features ", false),
+    ("Php classes and objects", false),
+    ("Predefined variabales, exceptions, interfaces and clasess", false),
+    ("Security: session, filesystem, database, error", false);
+   
+
+   INSERT INTO python (python_name, python_condition)
+VALUES ("How to start using Python", false),
+    ("Python language basics, objects and classes", false),
+    ("Problem solving with algorythyms and data strutures", false),
+    ("Python's popularirty as a programming lanaguange", false),
+    ("Executing a script in Python", false);
+   
+
+
+
