--- models/activities.js
@@ -1,9 +1,0 @@
-module.exports = function(sequelize, DataTypes) {
-    var Activities = sequelize.define("Activities", {
-      activity_name: DataTypes.STRING,
-      a_condition: DataTypes.BOOLEAN}, { 
-      timestamps: false,
-    });
-    return Activities;
-  };
-
