+++ models/php.js
@@ -1,0 +1,9 @@
+module.exports = function(sequelize, DataTypes) {
+    var Php = sequelize.define("Php", {
+      php_name: DataTypes.STRING,
+      php_condition: DataTypes.BOOLEAN}, { 
+      timestamps: false,
+    });
+    return Php;
+  };
+
