+++ models/python.js
@@ -1,0 +1,9 @@
+module.exports = function(sequelize, DataTypes) {
+    var Python = sequelize.define("Python", {
+      python_name: DataTypes.STRING,
+      python_condition: DataTypes.BOOLEAN}, { 
+      timestamps: false,
+    });
+    return Python;
+  };
+
