--- models/topics.js
@@ -1,13 +1,0 @@
-module.exports = function(sequelize, DataTypes) {
-  var Topics = sequelize.define(
-    "Topics",
-    {
-      topic_name: DataTypes.STRING,
-      t_condition: DataTypes.BOOLEAN,
-    },
-    {
-      timestamps: false,
-    }
-  );
-  return Topics;
-};
