+++ public/signup.html
@@ -200,7 +200,7 @@
     <script>
       const form = document.getElementById("contact-form");
 
-      const formEvent = form.addEventListener("click", (event) => {
+      const formEvent = form.addEventListener("submit", (event) => {
         event.preventDefault();
         let mail = new FormData(form);
         sendMail(mail);
