+++ routes/api-routes.js
@@ -1,6 +1,31 @@
 // Requiring our models and passport as we've configured it
 const db = require("../models");
 const passport = require("../config/passport");
+require("dotenv").config();
+const nodemailer = require("nodemailer");
+
+
+const transporter = nodemailer.createTransport({
+  // host: "smtp-mail.outlook.com",
+  // secureConnection: false,
+  // port: 587,
+  // logger: true,
+  // debug: true,
+
+  // host: "smtp.gmail.com",
+  // port: 465,
+  // secure: true,
+  service: "gmail",
+  auth: {
+    user: process.env.EMAIL,
+    pass: process.env.PASS,
+  },
+  tls: {
+    // do not fail on invalid certs
+    rejectUnauthorized: true,
+    ciphers: "SSLv3",
+  },
+});
 
 module.exports = function(app) {
   // Using the passport.authenticate middleware with our local strategy.
@@ -50,4 +75,85 @@ module.exports = function(app) {
       });
     }
   });
+
+  // ========== NODEMAILER ========== //
+// const transporter = nodemailer.createTransport({
+//   // host: "smtp-mail.outlook.com",
+//   // secureConnection: false,
+//   // port: 587,
+//   // logger: true,
+//   // debug: true,
+
+//   // host: "smtp.gmail.com",
+//   // port: 465,
+//   // secure: true,
+//   service: "gmail",
+//   auth: {
+//     user: process.env.EMAIL,
+//     pass: process.env.PASS,
+//   },
+//   tls: {
+//     // do not fail on invalid certs
+//     rejectUnauthorized: true,
+//     ciphers: "SSLv3",
+//   },
+// });
+
+// verify connection configuration
+transporter.verify(function(error, success) {
+  if (success) {
+    console.log("verification success");
+  } else {
+    console.log("error", error);
+  }
+});
+
+const mailOptions = {
+  from: "codebasicsonlineportal@gmail.com", // sender address (who sends)
+  to: 'feelthehousegroove@yahoo.com, ivan_e21@hotmail.com', // list of receivers (who receives)
+  subject: "Welcome to CODEBASICS.COM", // Subject line
+  text: "Hello world, this is the first email", // plaintext body
+};
+
+// send mail with defined transport object
+transporter.sendMail(mailOptions, function(error, info) {
+  if (error) {
+    return console.log(error);
+  }
+  console.log("Message sent: " + info.response);
+});
+
+app.post("/api/signup", (req, res) => {
+  db.User.create({
+    // email: req.body.email,
+    // password: req.body.password,
+    // firstName: req.body.firstName,
+    // lastName: req.body.lastName,
+    
+    name: req.body.name,
+    email: req.body.email,
+    message: req.body.message
+  })
+    .then(() => {
+      res.redirect(307, "/");
+      const mailOptions = {
+        from: "codebasicsonlineportal@gmail.com",
+        to: req.body.email,
+        subject: "Welcome to CODEBASICS.COM",
+        text: "Hello world, this is the first email",
+      };
+      transporter.sendMail(mailOptions, (err) => {
+        if (err) {
+          console.log("Error has occured");
+        } else {
+          console.log("Email Sent");
+        }
+      });
+    })
+    .catch((err) => {
+      res.status(401).json(err);
+    });
+});
+// ========== END NODEMAILER ========== //
+
 };
