+++ routes/html-routes.js
@@ -25,9 +25,9 @@ module.exports = function(app) {
   // Here we've add our isAuthenticated middleware to this route.
   // If a user who is not logged in tries to access this route they will be redirected to the signup page
   app.get("/members", isAuthenticated, (req, res) => {
-    db.Topics.findAll({}).then(function(dbTopics) {
-      db.Activities.findAll({}).then(function(dbActivities) { 
-           res.render("members", {Topics: dbTopics, Activities: dbActivities})
+    db.Html.findAll({}).then(function(dbHtml) {
+      db.Css.findAll({}).then(function(dbCss) { 
+           res.render("members", {Html: dbHtml, Css: dbCss})
       })
   });
   });
