+++ views/members.handlebars
@@ -60,20 +60,21 @@
 	{{!-- <div class="col-md-3"></div> --}}
 		<div class="col-md-3">
 			<h3>HTML:</h3>
-			<h4><a href="https://developer.mozilla.org/en-US/docs/Web/HTML">Resources</a></h4>
-			<ul id="unfinishedTopics">
+			<h4><a href="https://html.com/">Resources</a></h4>
+			<ol id="unfinishedTopics">
 				{{#each Topics}}
 
 				{{#unless this.dataValues.t_condition}}
-				<li>
-					<h5>{{this.dataValues.id}}.{{this.dataValues.topic_name}}
-						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-primary"
+				<li><p> 
+					<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-sm btn-primary"
 							data-id="{{this.dataValues.id}}"><i class="far fa-check-square"></i></button>
-					</h5>
+					 {{this.dataValues.topic_name}}
+						
+					</p>
 				</li>
 				{{/unless}}
 				{{/each}}
-			</ul>
+			</ol>
 
 			<form class="create-update-form" action="api/insertTopic" method="POST">
 				<div>
@@ -86,19 +87,19 @@
 		<div class="col-md-3">
 			<h3>CSS:</h3>
 			<h4><a href="https://devdocs.io/css/">Resources</a></h4>
-			<ul id="unfinishedActivities">
+			<ol id="unfinishedActivities">
 				{{#each Activities}}
 				{{#unless this.dataValues.a_condition}}
-				<li>
-					<h5>{{this.dataValues.id}}.{{this.dataValues.activity_name}}
-						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-primary"
-							data-id="{{this.dataValues.id}}">COMPLETE!</button>
-					</h5>
-
+				<li><p> 
+					<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-sm btn-primary"
+							data-id="{{this.dataValues.id}}"><i class="far fa-check-square"></i></button>
+					 {{this.dataValues.topic_name}}
+						
+					</p>
 				</li>
 				{{/unless}}
 				{{/each}}
-			</ul>
+			</ol>
 
 			<form class="create-update-form" action="api/insertTopic" method="POST">
 				<div>
@@ -108,7 +109,7 @@
 			</form>
 		</div>
 
-		<div class="col-md-3">
+		{{!-- <div class="col-md-3">
 			<h3>JavaScript:</h3>
 			<h4><a href="https://javascript.info/">Resources</a></h4>
 			<ul id="unfinishedActivities">
@@ -116,8 +117,8 @@
 				{{#unless this.dataValues.a_condition}}
 				<li>
 					<h5>{{this.dataValues.id}}.{{this.dataValues.activity_name}}
-						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-primary"
-							data-id="{{this.dataValues.id}}">COMPLETE!</button>
+						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-sm btn-primary"
+							data-id="{{this.dataValues.id}}"><i class="far fa-check-square"></i></button>
 					</h5>
 
 				</li>
@@ -133,7 +134,7 @@
 			</form>
 		</div>
 
-	</div>
+	</div> --}}
 	<br>
 
 	<div class="row justify-content-center">
@@ -171,19 +172,20 @@
 		<div class="col-md-3">
 			<h3>Python:</h3>
 			<h4><a href="https://wiki.python.org/moin/BeginnersGuide/Programmers">Resources</a></h4>
-			<ul id="unfinishedTopics">
+			<ol id="unfinishedTopics">
 				{{#each Topics}}
 
 				{{#unless this.dataValues.t_condition}}
-				<li>
-					<h5>{{this.dataValues.id}}.{{this.dataValues.topic_name}}
-						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-primary"
-							data-id="{{this.dataValues.id}}">COMPLETE!</button>
-					</h5>
+				<li><p> 
+					<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-sm btn-primary"
+							data-id="{{this.dataValues.id}}"><i class="far fa-check-square"></i></button>
+					 {{this.dataValues.topic_name}}
+						
+					</p>
 				</li>
 				{{/unless}}
 				{{/each}}
-			</ul>
+			</ol>
 
 			<form class="create-update-form" action="api/insertTopic" method="POST">
 				<div>
@@ -193,7 +195,7 @@
 			</form>
 		</div>
 
-		<div class="col-md-3">
+		{{!-- <div class="col-md-3">
 			<h3>JAVA:</h3>
 			<h4><a href="https://www.programiz.com/java-programming">Resources</a></h4>
 			<ul id="unfinishedActivities">
@@ -201,8 +203,8 @@
 				{{#unless this.dataValues.a_condition}}
 				<li>
 					<h5>{{this.dataValues.id}}.{{this.dataValues.activity_name}}
-						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-primary"
-							data-id="{{this.dataValues.id}}">COMPLETE!</button>
+						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-sm btn-primary"
+							data-id="{{this.dataValues.id}}"><i class="far fa-check-square"></i></button>
 					</h5>
 
 				</li>
@@ -216,24 +218,24 @@
 					<button class="btn btn-primary" type="submit">Add topic</button>
 				</div>
 			</form>
-		</div>
+		</div> --}}
 
 		<div class="col-md-3">
 			<h3>PHP:</h3>
 			<h4><a href="https://www.php.net/manual/en/">Resources</a></h4>
-			<ul id="unfinishedActivities">
+			<ol id="unfinishedActivities">
 				{{#each Activities}}
 				{{#unless this.dataValues.a_condition}}
-				<li>
-					<h5>{{this.dataValues.id}}.{{this.dataValues.activity_name}}
-						<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-primary"
-							data-id="{{this.dataValues.id}}">COMPLETE!</button>
-					</h5>
-
+				<li><p> 
+					<button type='submit' id={{this.dataValues.id}} class="complete-btn btn btn-sm btn-primary"
+							data-id="{{this.dataValues.id}}"><i class="far fa-check-square"></i></button>
+					 {{this.dataValues.topic_name}}
+						
+					</p>
 				</li>
 				{{/unless}}
 				{{/each}}
-			</ul>
+			</ol>
 
 			<form class="create-update-form" action="api/insertTopic" method="POST">
 				<div>
